backend/requirements.txt
fastapi
uvicorn
python-dotenv
PyGithub
python-multipart
cryptography
pydantic_settings
python-jose
httpx
pydantic